# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DigitalTicket API
  version: 0.1.0
  description: >-
    API multi-tenant para eventos, ingressos, pagamentos, check-in,
    relatórios, cupons e white-label. Valores monetários são expressos em
    centavos de BRL.
servers:
  - url: http://localhost:4000
    description: Desenvolvimento local
  - url: https://api.example.com
    description: Produção (configure o domínio real)
...
paths:
...
  /api/v1/check-in/catalog:
    get:
      tags: [Check-in]
      summary: Baixa catálogo de ingressos para cache offline e busca manual
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: eventId, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 2000, default: 500 } }
      responses:
        '200': { description: Ingressos com hash QR e status }
        '403': { $ref: '#/components/responses/Error' }
  /api/v1/check-in/stats:
    get:
      tags: [Check-in]
      summary: Retorna entradas versus total vendido
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: eventId, in: query, schema: { type: string } }
      responses:
        '200': { description: Estatísticas em tempo real }
  /api/v1/check-in/history:
    get:
      tags: [Check-in]
      summary: Lista o histórico do operador autenticado
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 100 } }
      responses:
        '200': { description: Leituras recentes }
  /api/v1/check-in/validate:
    post:
      tags: [Check-in]
      summary: Valida QR Code/código e registra entrada
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckInInput' }
      responses:
        '200': { description: Check-in aprovado }
        '409': { description: Ingresso já utilizado }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT retornado por /api/v1/auth/login
  responses:
    Error:
      description: Erro da API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        details: { type: object, additionalProperties: true }
...
    CheckInInput:
      type: object
      properties:
        signedToken: { type: string }
        checkInCode: { type: string }
        ticketId: { type: string }
...